+++ imio/dms/mail/migrations/migrate_to_3_1_6.py
@@ -1,12 +1,18 @@
 # -*- coding: utf-8 -*-
 from imio.dms.mail.migrations.migrate_to_3_1 import Migrate_To_3_1
 
+import logging
+
+
+logger = logging.getLogger("imio.dms.mail")
+
 
 class Migrate_To_3_1_6(Migrate_To_3_1):  # noqa
 
     def run_parts(self):
 
         if self.is_in_part("c"):
+            self.fix_localroles_json_quoting()
             # update settings
             self.runProfileSteps("imio.dms.mail", steps=["plone.app.registry"])